+++ week_three/lab/Activity2PayStub.java
@@ -1,13 +1,13 @@
-import java.util.Scanner; 
+import java.util.Scanner;
 
-/** * Activity1PayStub class is part of Lab 3 and 
- * creates a simple pay stub. 
- * 
- * @author Hunter_Wilson 
- * @version 9.12.2023 
- */ 
-public class Activity2PayStub 
-{ 
+/**
+ * Activity1PayStub class is part of Lab 3 and
+ * creates a simple pay stub.
+ *
+ * @author Hunter_Wilson
+ * @version 9.12.2023
+ */
+public class Activity2PayStub {
     public static final double OVERTIME_RATE = 1.5;
     public static final double SS_RATE = 0.1;
     public static final double TAX_RATE = 0.2;
@@ -18,19 +18,19 @@ public class Activity2PayStub
     private int hoursOvr;
     private double hourlyRat;
     private double regularPay;
-    private double overtimePayRate; 
+    private double overtimePayRate;
     private double overtimePay;
-    private double ssHolding; 
-    private double fedTax; 
+    private double ssHolding;
+    private double fedTax;
     private double netPay;
     private double grossPay;
+
     /**
      * It all starts with the main method.
      *
      * @param args command-line arguments (not used)
      */
-    public static void main(String[] args)
-    {
+    public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
 
         //Create an Activity2Paystub object
@@ -43,14 +43,14 @@ public class Activity2PayStub
         a2ps.printPayStub();
     }
 
-    /** * getInput gets input from keyboard.
-     * 
-     * @author Hunter_Wilson 
-     * @version 9.12.2023 
+    /**
+     * getInput gets input from keyboard.
+     *
      * @param scan is the scanner
-     */ 
-    public void getInput(Scanner scan) 
-    { 
+     * @author Hunter_Wilson
+     * @version 9.12.2023
+     */
+    public void getInput(Scanner scan) {
         System.out.print("Enter Employee Name: ");
         employeeName = scan.nextLine();
 
@@ -65,31 +65,31 @@ public class Activity2PayStub
 
         System.out.print("Enter Hourly Pay Rate: ");
         hourlyRat = scan.nextDouble();
-    } 
-    
-    /** * calculate, calculates the pay and such.
-     * 
-     * @author Hunter_Wilson 
-     * @version 9.12.2023 
-     */ 
-    public void calculate() 
-    { 
-        regularPay = (double) hours * (double) hourlyRat;
+    }
+
+    /**
+     * calculate, calculates the pay and such.
+     *
+     * @author Hunter_Wilson
+     * @version 9.12.2023
+     */
+    public void calculate() {
+        regularPay = (double) hours * hourlyRat;
         overtimePayRate = hourlyRat * OVERTIME_RATE;
         overtimePay = hoursOvr * overtimePayRate;
         grossPay = regularPay + overtimePay;
         ssHolding = grossPay * SS_RATE;
         fedTax = (grossPay - ssHolding) * TAX_RATE;
         netPay = (grossPay - ssHolding) - fedTax;
-    } 
+    }
 
-    /** * printPayStub, prints the paystub in terminal.
-     * 
-     * @author Hunter_Wilson 
-     * @version 9.12.2023 
+    /**
+     * printPayStub, prints the paystub in terminal.
+     *
+     * @author Hunter_Wilson
+     * @version 9.12.2023
      */
-    public void printPayStub() 
-    { 
+    public void printPayStub() {
         String ln1 = "________________________________";
         String ln2 = "Name: %-37s SSN: %-11s\n";
         String ln3 = "Regular Hours: %-8d Reg Rate: $%-8.2f Reg Pay: $%-8.2f\n";
@@ -109,5 +109,5 @@ public class Activity2PayStub
         System.out.printf(ln7, fedTax);
         System.out.printf(ln8, netPay);
         System.out.println(ln1 + ln1);
-    } 
+    }
 }
